增加点击头像，对其禁言，加关注，举报，拉黑功能

+++ MidnightArtifact/app/src/main/java/com/atman/wysq/model/response/ChatRoomMessageModel.java
@@ -30,7 +30,7 @@ public class ChatRoomMessageModel {
     private double sendTime;
     private int isAnchorImage;
     private int cmdType;
-    private int banChatUserId;
+    private long banChatUserId;
 
     public String getGiftName() {
         return giftName;
@@ -120,11 +120,11 @@ public class ChatRoomMessageModel {
         this.cmdType = cmdType;
     }
 
-    public int getBanChatUserId() {
+    public long getBanChatUserId() {
         return banChatUserId;
     }
 
-    public void setBanChatUserId(int banChatUserId) {
+    public void setBanChatUserId(long banChatUserId) {
         this.banChatUserId = banChatUserId;
     }
